move match deletion to deletion
in preparation for resusing the scheduler for selection deletion
